refactor: move setting-menu-modal to common
