+++ packages/web/src/views/pool/pool-increase-liquidity/components/increase-liquidity/increase-select-position/IncreaseAmount.tsx
@@ -1,11 +1,13 @@
-import IconAdd from "@components/common/icons/IconAdd";
-import IconSettings from "@components/common/icons/IconSettings";
-import TokenAmountInput from "@components/common/token-amount-input/TokenAmountInput";
-import SettingMenuModal from "@components/swap/setting-menu-modal/SettingMenuModal";
-import { TokenAmountInputModel } from "@hooks/token/use-token-amount-input";
-import { TokenModel } from "@models/token/token-model";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
+
+import IconAdd from "@components/common/icons/IconAdd";
+import IconSettings from "@components/common/icons/IconSettings";
+import SettingMenuModal from "@components/common/setting-menu-modal/SettingMenuModal";
+import TokenAmountInput from "@components/common/token-amount-input/TokenAmountInput";
+import { TokenAmountInputModel } from "@hooks/token/use-token-amount-input";
+import { TokenModel } from "@models/token/token-model";
+
 import { IncreaseSelectPositionWrapper } from "./IncreaseSelectPosition.styles";
 
 export interface IncreaseSelectPositionProps {
